Fix encoder handling of gvar in mixed mode.
In mixed mode patch generation special treatment is needed for gvar. The gvar table is normally augmented only by IFTB patches, so by default should not be included in the patches. However, in the case where the design space is expanded then the gvar table needs to be overwritten (so that glyphs in the base subset have updated gvar data). Non-base glyphs will be updated via IFTB patches. This updates encoder handling to add a special replacemnt gvar that only covers the base subset glyphs.